+++ experts/ExpertDessinPolygone.java
@@ -32,18 +32,18 @@ public class ExpertDessinPolygone extends ExpertDessin
 			fen.changerCouleur(tabStrings[1]);
 			for (int i = 2; i < tabStrings.length - 3; i = i+2) {
 				fen.dessinerLigne(
-						Integer.parseInt(tabStrings[i]),
-						Integer.parseInt(tabStrings[i+1]),
-						Integer.parseInt(tabStrings[i+2]),
-						Integer.parseInt(tabStrings[i+3])
+						(int)Double.parseDouble(tabStrings[i]),
+						(int)Double.parseDouble(tabStrings[i+1]),
+						(int)Double.parseDouble(tabStrings[i+2]),
+						(int)Double.parseDouble(tabStrings[i+3])
 				);
 			}
 			
 			fen.dessinerLigne(
-					Integer.parseInt(tabStrings[2]),
-					Integer.parseInt(tabStrings[3]),
-					Integer.parseInt(tabStrings[tabStrings.length - 2]),
-					Integer.parseInt(tabStrings[tabStrings.length - 1])
+					(int)Double.parseDouble(tabStrings[2]),
+					(int)Double.parseDouble(tabStrings[3]),
+					(int)Double.parseDouble(tabStrings[tabStrings.length - 2]),
+					(int)Double.parseDouble(tabStrings[tabStrings.length - 1])
 			);
 
 			System.out.println("c'est un polygone !");
